modify article_id on /api/comments for articles_id

+++ src/api/comments.js
@@ -5,7 +5,18 @@ const app       = require("../app.js");
 app.route("/api/comments/create")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
-        
+        if (typeof req.body.content !== "string" || req.body.content === "") {
+            res.status(503).send({ status: "Error", extra: "Le contenu de l'article est vide" });
+            return;
+        }
+        if (typeof req.body.author !== "string" || req.body.author === "") {
+            res.status(503).send({ status: "Error", extra: "L'auteur n'est pas reinsigné"  });
+            return;
+        }
+        if (typeof req.body.articles_id !== "string" || req.body.articles_id === "") {
+            res.status(503).send({ status: "Error", extra: "Vous devez rensigner un article id" });
+            return;
+        }
 
         const sqlConnection = mysql.createConnection(sqlConfig);
         sqlConnection.query(
@@ -26,8 +37,12 @@ app.route("/api/comments/create")
 app.route("/api/comments/delete")
     .get((req, res) => res.status(503).send({ status: "ERROR"}))
     .post((req, res) => {
-        const sqlConnection = mysql.createConnection(sqlConfig);
+        if (typeof req.body.id !== "string" || req.body.id === "") {
+            res.status(503).send({ status: "Error", extra: "Vous devez rensigner une Id" });
+            return;
+        }
 
+        const sqlConnection = mysql.createConnection(sqlConfig);
         sqlConnection.query(
             "DELETE FROM node_comments WHERE id = ?",
             [req.body.id],
@@ -45,13 +60,16 @@ app.route("/api/comments/delete")
 
 app.get("/api/comments", (req, res) => {
     const sqlConnection = mysql.createConnection(sqlConfig);
+
     sqlConnection.query(
-        "SELECT articles_id, content, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
+        "SELECT node_comments.id, articles_id, content, node_users.firstname AS authorFirstname, node_users.lastname AS authorLastname, created_at"
         + "  FROM node_comments"
         + "  LEFT JOIN node_users"
         + "  ON node_comments.author = node_users.id"
+        + "  WHERE articles_id = ?"
         + "  ORDER BY created_at DESC"
         + "  LIMIT 5;",
+        [req.query.id],
         (error, result) => {
             if (error) {
                 console.log(error.code);
@@ -76,7 +94,7 @@ app.get("/api/comment", (req, res) => {
         + "  LEFT JOIN node_users"
         + "  ON node_comments.author = node_users.id"
         + "  WHERE articles_id = ?"
-        + "  LIMIT 5;",
+        + "  LIMIT 1;",
         [ req.query.articles_id ],
         (error, result) => {
             if (error) {
@@ -93,3 +111,4 @@ app.get("/api/comment", (req, res) => {
         }
     );
 });
+
